Continued refactoring - Changed db_args from Vec<&str> to Vec<String> - Moved most debugger methods to debugger_functions - Expanded functionality of DebuggerState
